+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Fixed
 * Fix incorrect affects ordering by Impact and Resolution (`OSIDB-3480`)
 
+### Added
+* Support PURLs in affected components (`OSIDB-3412`)
+
 ## [2024.12.1]
 ### Fixed
 * Fix save all button behavior for affects table (`OSIDB-3664`)